perf: improve frontend load times and backend query performance
Frontend:
- Dynamic import Framer Motion + TypeAnimation on homepage (reduces initial JS ~3MB)
- Convert Geologica font from TTF (343KB) to WOFF2 (120KB) with font-display swap
- Remove console.log calls from theme provider in production

Backend:
- Add compound indexes: friendships.by_requester_status, events.by_creator_visibility
- Use compound indexes to filter at DB level instead of JS (friendships, events)
- Parallelize sequential DB reads with Promise.all in getUserEventsAndInvites,
  getDiscoverableEvents, getFriends, getMutualFriends

Closes #152, #153, #154, #155, #157
